Add logout button to layout component
