+++ frontend/src/pages/DashboardPage.tsx
@@ -1,6 +1,4 @@
 import React, { useEffect, useState } from 'react';
-import { useAuth } from '../context/AuthContext';
-import { useNavigate } from 'react-router-dom';
 import axios from 'axios'
 import Layout from '../components/Layout';
 
@@ -14,13 +12,7 @@ import Layout from '../components/Layout';
   }
 
 const DashboardPage: React.FC = () => {
-  const { logout } = useAuth();
-  const navigate = useNavigate();
 
-  const handleLogout = () => {
-    logout();
-    navigate('/login');
-  };
 
   const  [properties, setProperties] = useState<Property[]>([])
 
@@ -43,12 +35,6 @@ const DashboardPage: React.FC = () => {
         <div className="text-center space-y-4">
           <h1 className="text-3xl font-bold">Welcome to RentWise</h1>
           <p className="text-lg">You're successfully authenticated.</p>
-          <button
-            onClick={handleLogout}
-            className="px-4 py-2 bg-red-600 text-white rounded hover:bg-red-700 transition"
-          >
-            Logout
-          </button>
         </div>
       </div>
     </Layout>
